Add a "passthrough" variant of register assignment.
For sext/zext there are some cases where absolutely no code needs to be
generated in general. Previously we ended up with no-ops such as:

```
; %2: i64 = ...
...
; %3: ptr = zext %2
mov eax, eax
```

The CPU is, in general, clever enough to realise this is a no-op and
there is no performance penalty for it. However, it muddies the output,
and also makes it less obvious that there's the (good!) potential for
register aliasing: in some cases, we could use the same register to hold
the value of multiple SSA variables.

This commit doesn't go that far, but it makes clearer that it is at
least possible.