Detect and abort tracing if "early return" occurs.
If we start tracing a loop and the function-with-a-control-point returns
whilst tracing is ongoing, we are in dangerous territory. For quite a
while we assumed that "dangerous territory" meant (gongs of doom!) "UB".
Actually, we just hadn't created very good examples: when I did, it
turns out that an `unwrap` in trace_builder fails if early return
happens. We *could*, therefore, detect early returns there, but that
means that we've done a lot of work before realising this.

This commit instead has `mt.rs` detect early returns, and abort tracing,
as soon as we reasonably can. [Note, however, that if a user annotates
`return`s from the function, they could stop things even earlier. Is
this worth it? It is quite complicated to explain and, as we'll see in a
bit, it's not a very common case]

All that said, it turns out early returns aren't that common: you have
to start tracing on the same iteration that you "early return" from.
This is definitely possible -- indeed the two C tests in the PR do that,
one "normally" and one when side-tracing -- but it seems unlikely that
it will happen very often. So although we could force the user to help
us detect this even sooner, I'm not sure it's worth burdening them or us
with that responsibility right now: the cases where it would save
serious resources are borderline pathological. Indeed, the only such
case I can think of is: if you start tracing from the "root" control
point call frame, and then return to the main program (i.e. you will
never go back to the interpreter), then tracing is never turned off. One
day we can fix this, but I don't think we will encounter it very soon.