Optimise the body after peeling.
Right now optimising each instruction as its copied probably works OK,
but it's going to be fragile in the long-term. In particular, this puts
a *very* strong constraint on `iidx_map` that I suspect we'll forget.

This commit turns this into a stage two thing: first we peel, creating a
body that's virtually identical to the header. Once that's completely
done, we then run the optimiser over the complete body. This means that
optimising the header and body now run in conceptually the same manner:
it's a bit slower (though probably unmeasureably so!) but less likely
for us to shoot ourselves in the foot.