Add C tests for peeling.
Previously it was impractical for us to write a test for peeling:
limitations in the optimiser meant that you needed a complex set of
things to go right for peeling to end up having visibly good effects.
The previous commit to this has made it possible for us to write a small
test that makes the effect of peeling obvious: in this case we are able
to elide a guard in the body. Irritatingly, we have to have minor
variants for hwt/swt, because the latter inserts a store.