Unbreak `YKD_OPT=0`.
When adding support for loop peeling, we forgot that loops are only
peeled when optimisations are turned on. When they were turned off, we
generated code with no backjump e.g.:

```
; %294: ptr = load %293
mov rsi, [rdi+0x588]
; header_end [...]
; Deopt ID for guard 0
push rsi
mov rsi, 0x00
```

Notice no `jmp` between `header_end` and `Deopt`. That meant that we
fell through to whichever guard's deopt happened to come first, with
confusing results. For example, before this commit the new test
`ykd_opt_off` would print:

```
10
9
8
7: 7
```

even though the code shows this is clearly "impossible":

```rust
if (i < 7)
  printf("7: ");
```

This commit not only fixes that, but brings a bit more order to three
kinds of traces we can compile: "header only" traces (most tests, and
when `YKD_OPT=0`); "header and body" traces (with loop peeling); and
"side traces". A new `TraceKind` captures these, and allows us to
simplify quite a bit of code in the x64 backend, which no longer has to
guess what kind of trace it has.

Perhaps inevitably this also showed up a few tests where we'd muddled
header and body.