Forward register hints from header to body.
When we implemented peeling, we forgot -- it didn't even occur to me! --
to forward register hints from the header to the body. In essence, we
told the body where to put things, put not the header, so the register
allocator started the body in a different state to the one we wanted to
end up in.

This all worked, of course, because hints are just hints. However, the
lack of hints hugely increased the chances of needing to do lots of
register swaps at the end of the body.

Propagating the hints gets us another 12% or so on big_loop.