De-x64ise `VarLocation`.
This commit is a bit of a hack, but it starts to make it possible for us
to have more than one codegen active at any one time.

In order to do that, we have to make the `VarLocation` struct not hardcode
in x64 registers.

That means making `VarLocation` take a type parameter for the particular
codegen's register type. However, doing that causes a proliferation of
typing errors across the codebase. Fixing those would not be fun, and
a quick experiment suggests wouldn't even be a good idea: they just make
a lot of the code harder to read.

Instead, what this commit does is use a couple of strategically placed
`typedef`s of the form:

```
typedef VarLocation = reg_alloc::VarLocation<Register>;
```

That means that referring to just `VarLocation` is correct at nearly all
the points it ever was.

Note: this commit does not finish off the de-x64isation of jitc_yk, but
it does make the future path to doing so, and probable solutions,
clearer.