Pull side-trace codegen-specific stuff out of the JIT module.
This commit is, again, a bit of a hack, but it still makes things a bit
better than before. In essence, before this commit JIT modules contained
`yksmp::Location`s for "normal" traces and `VarLocation`s for side
traces. The former are codegen agnostic: the latter are codegen
specific.

This commit hoiks the `VarLocation`s out of the JIT mod and puts them in
`TraceKind::Sidetrace`. Now that would end up having us add a `Register`
type parameter everywhere, which is too much churn, so for now
`TraceKind::Sidetrace` stores an `Arc<SidetraceInfo>` i.e. we store it
as a trait object. We then have to downcast to `YkSideTraceInfo` in lots
of places. This is pretty evil, because we're using dynamic typing where
we should be using static typing. However, overall, this is the lesser
of the evils.

By the end of this commit, it's clearer (even if not perfectly so) the
relationship between trace inputs and side-trace outputs. We can do
better in the future, and maybe we'll want to do so, but this does feel
like a nudge in the right direction.