Introduce `assign_regs` to assign GP and FP registers in one go.
Mostly one can assign GP/FP regs in isolation. However, because x64
requires some FP operations to use a temporary GP register, we currently
find ourselves in the uncomfortable position where we have to `push` a
GP register in order to be certain that we aren't accidentally
clobbering something.

This commit introduces a new function in the API `assign_regs` which
specifies FP and GP registers in one go. Although we don't take
advantage of this yet, it will allow us (one day!) to safely deal with
the "I need a temporary GP register" case.