Remove intermediate functions in the register allocator.
Some queries the codegen and register allocator need to make are
"reverse analysis" queries. This commit makes clearer that these are
queries on `rev_an` rather than the register allocator.

Doing this means allowing another public(ish) field, which looks ugly,
but at least makes clearer what the "real" structure is. I increasingly
think we need to think of a "proper" fix for this, but our existing
structure is unfortunate in several different places, and I can't
immediately see an easy migration strategy.