Record (def, use) pairs for instructions during reverse analysis.
Previously the reverse analyser only recorded whether an instruction's
output value was used. This commit records which (later!) instructions
use an instruction's output value.

For the time being, we only use this to make a slightly better heuristic
about "what to spill". Previously we used the (fairly terrible) proxy of
"spill the value which lives the longest". Now we use the heuristic
"spill the value which is not used for the longest time". On small
examples, these two often coincide, but on bigger examples they
definitely won't.