When we need to find a spare register, only spill if we have to.
When we've run out of registers, we have to chose a register to "spill".
This commit recognises that we often don't have to spill anything,
either because the register contains a constant or because the register
is duplicating an already-spilled value. We thus prefer not to actually
spill (i.e. store!) anything if one of those cases happens, only
spilling if we really have to.