Don't consider `Param`s as removable instructions.
`Param`s are not normal instructions: they're effectively declarations of
the state of the system at the point that a trace is about to start.
Previously we accidentally treated these in two different ways:

  1. In "full, normal" compilation mode, they were implicitly never
     removed because they appeared in `TraceHeaderStart` and
     `TraceBodyStart`.

  2. In tests, they could be removed if they weren't "used".

This commit makes `Param`s never removable by treating them as "internal
instructions", so tests in (2) now observe the same behaviour as (1).