Spill registers at the start of the trace if possible.
When a trace starts, we typically have lots of `Param`s that appear in
guards and `TraceHeaderEnd`/`TraceBodyEnd` but nowhere else. These
registers then gum up the register allocator, making our spill choices
poor -- even though we don't typically use them much.

There are various ways we might be able to tackle this: this commit does
the simplest thing, which is to spill such registers at the very start
of the trace.