Add canonicalisation tests.
I noticed (because I made a mistake!) that we weren't testing the
canonicalisation cases properly. This commit looks like it changes much
more than it really does: basically we add a comment to the main
optimiser code which causes lots of indenting and churn without any
meaningful changes. The meat of the commit is the test cases at the end.