Expand the size of `InstIdx`.
Although `InstIdx` looks to be 16 bits big, `PackedOperand` forces it to
be 15 bits and that just isn't enough (32,767) for some of the traces we
now see, particularly after peeling.

Probably 19 bits would be big enough, but right now that level of bit
fiddling isn't worth our effort. Instead, this commit brute forces
`InstIdx` to 32 bits. That also forces `ConstIdx` to be 32 bits, to keep
code churn to a minimum.