Don't give up on a location the first time compiling fails.
This comment in the code was partly wrong:

```
// FIXME: We will endless retry tracing this
// [HotLocation]. There should be a `Counting`
// state.
```

In fact, we wouldn't endlessly retry: the `HotLocation` would get stuck
in the `Compiling` state. Still, the end effect was the same as the
comment suggested!

This commit allows `HotLocation`s to transition back to the `Counting`
state so that we can retry. That requires allowing `HotLocations` to
also be in the `Counting` state: in other words, we can "count" both
with and without a `HotLocation`.