Track nested meta-tracing states.
Previously we -- well, this one is entirely my fault, so "I"! -- tracked
per-thread meta-tracing state as a single `MTThreadState` that we updates as
necessary. This doesn't work when we have nested execution, tracing and
the like, as the bug in #1571 highlighted.

This commit moves us to a stack of `MTThreadState`s. The basic idea is
that the stack always has at least one element: `Interpreting`. As we go
through other states, we push / pop as appropriate. The test below (from
Edd and Lukas) fails on `master` but is fixed by this commit.

The implementation is a bit more awkward than one might hope as naive implementations either:

  1. Spread lots of knowledge about the stack around the code. That's a
     disaster waiting to happen.
  2. Run into borrow checker problems.

This commit gets around this in two phases:

  1. We pass closures to `MTThread` which can peek at the current
     `MTThreadState` (which isn't `Copy`!).
  2. Those closures return the "things" we need to update outside that
     context.

This is a bit awkward, and perhaps there's a better API, but this one is
at least safe.

Co-authored-by: Edd Barrett <vext01@gmail.com>
Co-authored-by: Lukas Diekmann <lukas.diekmann@gmail.com>