Catch another case of problematic control flow for tracing.
This change catches the case where we exit the function that started
tracing, but re-enter it later and stop tracing. This is problematic for
tracing, as evidenced by trace builder crashing with variable mapping
problems (it doesn't know about variables in the parent function).

We already have a check in place that was designed to detect "exiting
from the function that started tracing", but it's not sufficient for the
above scenario: it checks that the frame addresses at the time we start
and stop tracing are the same. Here they are the same.

To fix this, while building the trace, we check if a `Inst::Ret` would
cause the trace builder's frame stack to become empty.

It's still worth keeping the old check around, since they detect a
subset of the "bad returns" earlier than this new check.