Abort tracing as soon as the control point notices we're in a different frame.
Previously we detected "we're in a different frame" when we wanted to
stop tracing. This commit changes this so that the control point checks
this at all points a thread is tracing. This means that we can spot
problematic traces as they're ongoing, and abort them early.
Entertainingly, this allows us to simplify some surprisingly hairy code.