Stop tracing when we start unrolling.
Though it may not be obvious that this is what it's doing, this commit
stops us tracing undesirable things, notably when we start unrolling
inner loops.

Consider this example program:

```python
for x in range(...):
  if x > ...:
    for y in range(...):
```

The outer loop can get hot before the inner loop, so we start tracing
the outer loop -- which is fine. But when we trace, we may go around the
inner loop an unbounded number of times, endlessly unrolling the inner
loop. This can lead to gigantic, and "unstable" traces -- because we've
unrolled a specific number of iterations, there's a very high chance
that we'll see a different number of iterations next time and encounter
a guard on the first "iteration" of the gigantic-loop. This then leads
to us creating large numbers of essentially identical side-traces,
hugely bloating the system. Using havlak as an example this (combined
with a couple of other things) leads to ~10x more traces being created
than we would expect.

To avoid this happening, we have to track what `Location`s we've seen
while tracing. There are some potentially cunning things we could do
here if/when performance becomes an issue, but for now we force every
`Location` we see when tracing to become a `HotTracing`. That gives us a
stable ID (the `HotLocation`'s `Mutex`) which we can use as a reliable
proxy for "have we seen this `Location` before?"