Introduce "null" `YkLocation`s.
Previously we allowed users to pass `NULL` to `yk_mt_control_point` to
signify "there's nothing to trace at this point". However that meant
that most interpreters would need to maintain two structures for
locations: one to say "is there anything to trace at this point?" and
another for "this is the YkLocation for this point, if one is needed".
This was messy, and easy to misunderstand.

This commit no longer allows passing `NULL` to `yk_mt_control_point`,
and instead introduces "null"-equivalent `YkLocation`s which are created
by `Location::null()` or `yk_location_null()`. In essence, these tell
`yk_mt_control_point` that there's nothing to do at this point in the
high-level program, allowing us to compress the two structures we needed
previously into one.

Is this a perfect API? Probably not. We might be able to do better
again, but this is a useful improvement already, and has allowed me to
meaningfully improve yklua.

Note: because this is a breaking change, I have temporarily had to point
`.buildbot.sh` at my yklua fork. As soon as this merges, we can update
yklua, and then have `.buildbot.sh` point back to the main yklua
repository.