Force truncate to truncate.
Previously we incorrectly assumed that zero extending the input was
sufficient, but that's clearly not the case: a zero extended i32 when
truncated to an i16 can still have bits set that it shouldn't.

We could fix this by setting the output of truncation to be `Undefined`
which would lead to lazy zero extension. Instead, I've chosen, without
putting huge thought into the consequences of it, to eagerly zero extend in
`cg_trunc`. Time will tell if that's the best way of doing it or not!