Make trace IDs unique and hidden in an opaque type.
We will soon need to uniquely identify traces. We (well... I) were once
paranoid about a `u64` wrapping, but (a) we're not very likely to see
that happen (cough) (b) if it does, we can change to a `u128` or `u256`
or whatever other type we might need.

This commit thus does the following:

  1. Compiled trace IDs are now guaranteed unique. [Unless a `u64`
     overflows, at which point we predictably `panic`.]
  2. We wrap those in a new opaque struct `CompiledTraceId` so that we
     can change this later in a way that would make updating the
     relevant places easier.