Give traces a unique identifier.
Previously we had two "identifier" notions of "what trace is executing":

  1. A `u64` which was "semi-unique" because we worried about the `u64`
     wrapping.
  2. A pair `(root trace, [guard failure sequence])`.

We used notion (1) for debugging and notion (2) for deopt. Technically
only the latter was considered "exact". The problem with notion (2) is
that it requires unbounded memory (consider deeply nested side-traces).

This commit, in essence, removes notion (2) and settles on notion (1) as
the unique way of identifying a trace. In case a `u64` ever turns out
not to give us enough identifiers, I've introduced a `CompiledTraceId`,
which we can easily change to a `u128` or whatnot. I do not think this
is a realistic worry in the foreseeable future.

As such this commit is "just" a tidy up. However, it makes some other
deeper changes we will probably want feasible.