Prefer `Temporary` over `clobber_reg`.
`Temporary` will often have a spare register to hand out, and not need
to spill anything; `clobber_reg` will more often have to spill.

Now, we could do better here, because the register we're asking for only
needs to be unused *after* the inputs are processed. But, still, this is
better than nothing.