When writing jump vars, use a spare register if one is available.
Previously we did a `push rax; use rax; pop rax` dance for each spill we
needed to move -- for long traces there are quite a lot of these! One
thing we could do is do the `push/pop rax` dance once no matter how many
spills we use but, in general, we will find we have a spare register
lurking around that means we don't even have to do that.

This commit does just that: it looks for an unused register and uses
that as the scratch register.