Simplify an uncommon place.
This additional code is pointless in most cases and -- particularly if
we hit the `swap` case -- can be quite slow. Right now, this entire loop
is somewhat suboptimal, in the sense that it doesn't guarantee to find
the minimal set of register moves. We can definitely do better, but it's
more work than I want to think about right now.