Try and move, rather than spill, registers if possible.
Sometimes we have values in registers that we're going to need in the
future, but they're in registers that we need to use *right now*.
Previously, we spilled them immediately. This commit tries, when
possible, to move those values to another register. There is more we
could do to optimise this, but this still tends to reduce the number of
spills and unspills, so it's already a win.