Make the JIT parser hand out registers in Intel, not Dwarf, ordering.
Everyone and their dog knows what order Intel registers are stored in --
except, for reasons that I can't fathom (there must be a reason, surely?
but I don't know what it might be!), DWARF.

Unfortunately, because we use `yksmp::Location` to represent some
things, and that uses DWARF numbering, because stackmaps use DWARF
numbering, the confusion of DWARF numbering percolates up to our test
suite. There is no advantage to this and, indeed, it means I have to
have a permanent tab open with the DWARF -> everyone-else mapping.

This commit changes the JIT IR parser so that it hands out registers in
DWARF ordering. This makes my pea brain, at least, a little happier.