Don't force JIT IR to specify which number parameter it is.
Previously we forced users to write this in JIT IR:

```
%0: i64 = param 0
%1: i64 = param 1
```

even though the two numbers (`%0` and `0`) must "match" and the second
is therefore pointless.

This commit removes this necessity, and also makes it possible for us in
the future to have JIT IR with non-register parameters. For now the
above becomes:

```
%0: i64 = param reg
%1: i64 = param reg
```

As and when we need it we could add `param direct` etc. Until then, I
will be happy not to be told that I've got `param`'s index wrong again!