Use a definitely-spare register in `write_jump_vars`.
The way we worked out that we had a spare register was completely
incorrect, as it didn't respect the current state of the register
allocator (indeed, it always returned `RAX`). In particular, that meant
that we could use a "spare register" that still contained a value that
we needed to use to write jump vars! This only affected side traces: I
spotted this when finding that we could sometimes go wrong after a
sidetrace had completed a run, because we would write one random value.

This commit asks the register allocator for an empty register and uses
that instead. The `assert` added in this commit is our guarantee that
we don't reintroduce this bug in the future.

Note that `find_empty_gp_reg` is necessarily a slightly dangerous
function, but there's no better way to do it than this. So far I have
not been able to come up with a situation where we don't have an empty
register: if we do encounter that, we'll have to go back to the
`push`/`pop` dance.