Continue splitting the register allocator up into more testable chunks.
This commit splits the register allocator into two major parts:

  1. A non-mutating part which decides what registers to allocate.
  2. A mutating part which generates code.

Previously these were intermingled together in a way that was hard to
understand and impossible to test. Happily the second part is now tiny:
all of the hard stuff is in the first part.

This commit also takes a section of that first part -- the bit about
which I was least confident -- and makes it entirely stateless in
`reg_copies_to_actions` (i.e. it doesn't even read state), making it
much more easily testable in a unit test setting.

In essence (and the comments in the code explain this, I hope, in more
detail) the register allocator now works in terms of "actions": that is,
we gradually build up a sequence of things we will do in the future
(i.e. code we will generate), but we don't do them yet.

Happily this commit also makes the allocator simpler overall, despite
generating code that is at least as good as we did before.

There is more we need to do here, but this is a meaningful step in the
right direction.