Further optimise `PtrAdd`.
In essence this follows chains of `PtrAdd`s, collapsing them down where
possible. This allows some `PtrAdd`s to be optimised away completely, as
CSE then recognises them as being equivalent to previous instructions.

I've implemented this optimisation many times, but because it keeps
variables alive for longer, it played badly with the register allocator.
With the various changes we've made to that, it now works much better,
shaving more time off big_loop and friends.