Perform all `mprotect` operations in one go.
Previously we called `mprotect` once per patch, which when we have lots
of guards is inefficient. This commit first works out the span of
addresses that the call covers, then runs `mprotect` once.

While I'm here, slightly simplify a few factors, including the
realisation that `SeqCst` is not necessary since we now do a single
`mprotect` and a weaker `fence` then achieves the same effect of
ensuring that other threads using a lock will see all the same writes as
this thread.