Remove unused `guards` field.
Since 1d342f1f80727aac4eef6ba2a8f0db35cabe5ea8, we have not tracked the
chain of guard indexes. However, because of the convoluted path we take
to storing these, rustc couldn't identify this field as being unused.
Indeed, it took me a while to realise that we definitely weren't using
it!