Minimally move register states to allowing multiple instructions.
The fundamental change here is to move `RegState::FromInst(InstIdx,
..)` to `RegState::FromInst(Vec<InstIdx>, ..)` in anticipation of
allowing multiple instructions' values to reside in a single register.

As such this commit aims to be more-or-less no-op, but it isn't quite:
in particular we *have* to change the code for spilling an `i1` to avoid
borrow checker problems. Happily, the changed code is better.
Previously we did this horror:

```rust
let old_ext =
    self.align_extensions(asm, reg, RegExtension::ZeroExtended);
dynasm!(asm; mov BYTE [rbp - off], Rb(reg.code()));
self.align_extensions(asm, reg, old_ext);
```

i.e. we temporarily mutated an instruction's extensions. That clearly
can't work if you have multiple instructions of a different bit width in
a register (and the borrow checker now complains, anyway). So we move
that code to this:

```rust
if *ext == RegExtension::ZeroExtended {
    dynasm!(asm; mov BYTE [rbp - off], Rb(reg.code()));
} else {
    dynasm!(asm
        ; bt Rd(reg.code()), 0
        ; setc BYTE [rbp - off]
    );
}
```

This does mutate the CPU flags -- and we now have to be clear that the
register allocator can do that! -- but it doesn't mutate the value in a
register.