Introduce new `AlignExtension` register constraint.
This specifically targets `cg_[sz]ext`, both of which need to force
extension alignment, but which do not necessarily need to also clobber
registers.

This commit is close to a no-op: it introduces `AlignExtension`, but
makes it more-or-less the same as `InputOutput`. In other words, this
commit doesn't take advantage of the new extesion, so we can review the
minimal change in isolation. We'll take advantage of it in a future
commit.